rename character data to match item data
