+++ module/character-data.mjs
