+++ system.json
@@ -5,7 +5,7 @@
     "version": "0.0.1",
     "compatibility": {
         "minimum": 12,
-        "verified": 12.328
+        "verified": "12.330"
     },
     "authors": [
         {
